+++ pychron/envisage/tasks/base_task.py
@@ -42,7 +42,7 @@ from pychron.envisage.tasks.actions import GenericSaveAction, GenericSaveAsActio
     GenericFindAction, RaiseAction, RaiseUIAction, ResetLayoutAction, \
     MinimizeAction, PositionAction, IssueAction, CloseAction, CloseOthersAction, AboutAction, OpenAdditionalWindow, \
     NoteAction, RestartAction, DocumentationAction, CopyPreferencesAction, SwitchUserAction, KeyBindingsAction, \
-    ChangeLogAction
+    ChangeLogAction, StartupTestsAction
 from pychron.loggable import Loggable
 
 
@@ -412,6 +412,7 @@ class BaseTask(Task, Loggable, PreferenceMixin):
             RestartAction(),
             KeyBindingsAction(),
             SwitchUserAction(),
+            StartupTestsAction(),
             id='help.menu',
             name='Help')
         return menu